Extract marker analysis helpers
