+++ js/lab-context.js
@@ -4,7 +4,8 @@ import { state } from './state.js';
 import { SBM_2015_THRESHOLDS, getEMFSeverity } from './schema.js';
 import { getStatus, hasCardContent, hashString, isDebugMode } from './utils.js';
 import { formatTime } from './theme.js';
-import { getActiveData, getEffectiveRangeForDate, getLatestValueIndex, getAllFlaggedMarkers } from './data.js';
+import { getActiveData } from './data.js';
+import { getAllFlaggedMarkers, getEffectiveRangeForDate, getLatestValueIndex } from './marker-analysis.js';
 import { getProfileLocation, getLatitudeFromLocation } from './profile.js';
 import { getBloodDrawPhases, getNextBestDrawDate, detectPerimenopausePattern, detectCycleIronAlerts } from './cycle.js';
 import { scanSupplementsForWarnings, humanizeEffect } from './supplement-warnings.js';
