+++ tests/test-correctness-phase2.js
@@ -148,6 +148,7 @@ const pwaAppShellAssets = [
   '/js/light-tool-camera-modals.js',
   '/js/light-tools.js',
   '/js/modal-lifecycle.js',
+  '/js/marker-analysis.js',
   '/js/sun.js',
   '/js/sun-active-session.js',
   '/js/sun-session-model.js',
